lab1/consumer/requirements.txt
azure-storage-file-datalake==12.14.0
azure-eventhub==5.11.4
requests==2.31.0
